+++ rialto/jobs/__init__.py
@@ -12,4 +12,4 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from rialto.jobs.decorators import datasource, job
+from rialto.jobs.decorators import config, datasource, job