+++ tests/jobs/test_job/test_job.py
@@ -11,9 +11,12 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+from rialto.jobs.decorators import config, datasource, job
 
 
-from rialto.jobs.decorators import datasource, job
+@config
+def custom_config():
+    return "config_return"
 
 
 @datasource